convert functions into methods
